fix: move table to own package
